[#26] Prevent potential buffer overflow in sizebuf
The size of the data type in the format specifier for the sprintf call
for printing to sizebuf can potentially print more characters than can
be held in the buffer. UINT64_MAX is 18446744073709551615, which is 20
characters. The size of sizebuf has been increased to 22 to hold the
maximum value, plus another character for the unit in most cases, plus
another for the null terminator.

This situation is not possible due to the size checks which occur in
order to get into the code block, but this will prevent the potential
buffer overflow.